requirements.txt
annotated-types==0.5.0
anyio==3.7.1
async-timeout==4.0.3
black==23.9.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
cryptography==41.0.4
ecdsa==0.18.0
fastapi==0.103.1
filelock==3.12.4
flake8==6.1.0
h11==0.14.0
httpcore==0.18.0
httpx==0.25.0
idna==3.4
iniconfig==2.0.0
isort==5.12.0
mccabe==0.7.0
mypy==1.5.1
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.2
platformdirs==3.10.0
pluggy==1.3.0
pyasn1==0.5.0
pycodestyle==2.11.0
pycparser==2.21
pydantic==2.3.0
pydantic_core==2.6.3
pyflakes==3.1.0
pytest==7.4.2
python-jose==3.3.0
python-on-whales==0.64.3
redis==5.0.0
requests==2.31.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.21
starlette==0.27.0
tqdm==4.66.1
typer==0.9.0
typing_extensions==4.8.0
urllib3==2.0.5
uvicorn==0.23.2
websockets==11.0.3
